add pipes

+++ main.c
@@ -26,7 +26,7 @@ int main(int argc, char **argv) {
 
   // Parse args.
   char *cmd_str = NULL;
-  int sleep_time = 0;
+  unsigned int sleep_time = 0;
 
   for (int i = 1; i < argc; i++) {
     if (strcmp(argv[i], "-c") == 0) {
@@ -34,7 +34,7 @@ int main(int argc, char **argv) {
       cmd_str = argv[i];
     } else if (strcmp(argv[i], "--sleep") == 0) {
       i++;
-      sleep_time = atoi(argv[i]);
+      sleep_time = (unsigned int)atoi(argv[i]);
     } else {
       giveup("unknown arg %s", argv[i]);
     }
